src/pyproject.toml
[build-system]
requires = [ "poetry>=1.0.5",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "flask-template"
version = "0.1.0"
description = ""
authors = [ "Appcookies",]
[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
default = true

[tool.yapf]
based_on_style = "google"
coalesce_brackets = true
dedent_closing_brackets = true

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true

[tool.poetry.dependencies]
aio-kong = "^0.7.0"
cent = "^3.0.1"
click = "^8.0.1"
confluent-kafka = "^1.6.0"
flask = "^2.0.1"
flask-babel = "^2.0.0"
gunicorn = "^20.0.4"
httpx = "^0.19.0"
marshmallow = "^3.5.0"
opentelemetry-api = "^1.3.0"
opentelemetry-exporter-jaeger = "^1.3.0"
opentelemetry-instrumentation-flask = "^0.22b0"
opentelemetry-sdk = "^1.3.0"
poetry = "^1.1.8"
pyjwt = "^1.7.1"
python = "^3.9"
raven = "^6.10.0"
redis-py-cluster = "^2.0.0"
ujson = "^1.35"

[tool.poetry.dev-dependencies]
a2wsgi = "^1.4.0"
black = "^21.11b1"
coverage = "^4.5"
flask-debugtoolbar = "^0.10.1"
isort = "^5.10.1"
pre-commit = "^2.15.0"
pytest-flask = "^0.15.0"
"ruamel.yaml" = "^0.16.7"
toml = "^0.10.2"
websockets = "^9.1"
yapf = "^0.31.0"
